Extract image-overlay frame source
